+++ src/ModelObserver.php
@@ -7,11 +7,11 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 class ModelObserver
 {
     /**
-     * Only dispatch the observer's events after all database transactions have committed.
+     * Indicates if Scout will dispatch the observer's events after all database transactions have committed.
      *
      * @var bool
      */
-    public $afterCommit = true;
+    public $afterCommit;
 
     /**
      * The class names that syncing is disabled for.
@@ -19,6 +19,16 @@ class ModelObserver
      * @var array
      */
     protected static $syncingDisabledFor = [];
+
+    /**
+     * Create a new observer instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->afterCommit = config('scout.after_commit', false);
+    }
 
     /**
      * Enable syncing for the given class.